+++ lib/convert.h
@@ -64,7 +64,7 @@ string_t convert_number_to_string(const long long integer);
  * @return string_t
  * @since v1.0.0
  */
-string_t convert_number_from_base_10_to_base(unsigned long long number, unsigned int base);
+string_t convert_number_from_base_10_to_base(unsigned long long number, unsigned long base);
 
 /**
  * @brief Convert a number with a specific base to a number base 10.
@@ -74,7 +74,7 @@ string_t convert_number_from_base_10_to_base(unsigned long long number, unsigned
  * @return int
  * @since v1.0.0
  */
-int convert_number_from_base_to_base_10(string_t number, unsigned int base);
+unsigned long convert_number_from_base_to_base_10(string_t number, unsigned long base);
 
 /**
  * @brief Convert a number with a specific base to a number of specific base.
@@ -85,6 +85,6 @@ int convert_number_from_base_to_base_10(string_t number, unsigned int base);
  * @return string_t
  * @since v1.0.0
  */
-string_t convert_number_from_base_to_another(string_t number, int base_from, int base_target);
+string_t convert_number_from_base_to_another(string_t number, unsigned long base_from, unsigned long base_target);
 
 #endif
